+++ src/SolutionInfo.cs
@@ -6,8 +6,8 @@
 using System.Reflection;
 
 [assembly: AssemblyProduct("Wyam")]
-[assembly: AssemblyVersion("0.12.3")]
-[assembly: AssemblyFileVersion("0.12.3")]
-[assembly: AssemblyInformationalVersion("0.12.3-beta")]
+[assembly: AssemblyVersion("0.12.4")]
+[assembly: AssemblyFileVersion("0.12.4")]
+[assembly: AssemblyInformationalVersion("0.12.4-beta")]
 [assembly: AssemblyCopyright("Copyright © Wyam Contributors")]