+++ src/Wyam.Configuration/NuGet/PackageInstaller.cs
@@ -1,12 +1,13 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using System.Text;
 using System.Threading.Tasks;
 using NuGet.Configuration;
 using NuGet.Frameworks;
 using NuGet.PackageManagement;
-using NuGet.ProjectManagement;
 using NuGet.Protocol.Core.Types;
 using Wyam.Common.IO;
 using Wyam.Common.Tracing;
@@ -17,53 +18,51 @@ namespace Wyam.Configuration.NuGet
 {
     public class PackageInstaller
     {
-        private readonly List<PackageSource> _packageSources = new List<PackageSource>
-        {
-            new PackageSource("https://api.nuget.org/v3/index.json")
-        };
         private readonly List<Package> _packages = new List<Package>();
+        private readonly NuGetLogger _logger = new NuGetLogger();
         private readonly IFileSystem _fileSystem;
-        private readonly WyamFolderNuGetProject _nuGetProject;
+        private readonly AssemblyLoader _assemblyLoader;
+        private readonly NuGetFramework _currentFramework;
+        private readonly ISettings _settings;
+        private readonly SourceRepositoryProvider _sourceRepositories;
+
         private DirectoryPath _packagesPath = "packages";
 
         internal PackageInstaller(IFileSystem fileSystem, AssemblyLoader assemblyLoader)
         {
             _fileSystem = fileSystem;
+            _assemblyLoader = assemblyLoader;
+            _currentFramework = GetCurrentFramework();
+            _settings = Settings.LoadDefaultSettings(fileSystem.RootPath.FullPath, null, new MachineWideSettings());
+            _sourceRepositories = new SourceRepositoryProvider(_settings);
+        }
 
-            // Get the current framework
+        private NuGetFramework GetCurrentFramework()
+        {
             string frameworkName = Assembly.GetExecutingAssembly().GetCustomAttributes(true)
                 .OfType<System.Runtime.Versioning.TargetFrameworkAttribute>()
                 .Select(x => x.FrameworkName)
                 .FirstOrDefault();
-            CurrentFramework = frameworkName == null
+            return frameworkName == null
                 ? NuGetFramework.AnyFramework
                 : NuGetFramework.ParseFrameworkName(frameworkName, new DefaultFrameworkNameProvider());
+        }
+        
+        // Note that sources are searched first at index 0, then index 1, and so on until a match is found
+        public void AddPackageSource(string packageSource) => _sourceRepositories.AddDefaultRepository(packageSource);
 
-            Settings = global::NuGet.Configuration.Settings.LoadDefaultSettings(_fileSystem.RootPath.FullPath, null, new MachineWideSettings());
-            IPackageSourceProvider packageSourceProvider = new PackageSourceProvider(Settings);
-            SourceRepositoryProvider = new WyamSourceRepositoryProvider(packageSourceProvider);
-            _nuGetProject = new WyamFolderNuGetProject(fileSystem, assemblyLoader, CurrentFramework, GetAbsolutePackagesPath().FullPath);
-            PackageManager = new NuGetPackageManager(SourceRepositoryProvider, Settings, GetAbsolutePackagesPath().FullPath)
-            {
-                PackagesFolderNuGetProject = _nuGetProject
-            };
+        public void AddPackage(string packageId, IEnumerable<string> packageSources, string versionRange,
+            bool getLatest, bool allowPrereleaseVersions, bool allowUnlisted, bool exclusive)
+        {
+            _packages.Add(new Package(_currentFramework, packageId, packageSources?.Select(_sourceRepositories.CreateRepository).ToList(),
+                versionRange, getLatest, allowPrereleaseVersions, allowUnlisted, exclusive));
         }
 
-        internal NuGetLogger Logger { get; } = new NuGetLogger();
-
-        internal INuGetProjectContext ProjectContext { get; } = new WyamProjectContext();
-
-        private ISettings Settings { get; }
-
-        private ISourceRepositoryProvider SourceRepositoryProvider { get; }
-
-        internal NuGetPackageManager PackageManager { get; }
-
-        internal NuGetFramework CurrentFramework { get; }
-
         public bool UpdatePackages { get; set; }
-        
-        public bool UseLocal { get; set; }
+
+        public bool UseLocalPackagesFolder { get; set; }
+
+        public bool UseGlobalPackageSources { get; set; }
 
         public DirectoryPath PackagesPath
         {
@@ -81,49 +80,41 @@ namespace Wyam.Configuration.NuGet
         private DirectoryPath GetAbsolutePackagesPath()
         {
             DirectoryPath packagesPath = _fileSystem.RootPath.Combine(PackagesPath).Collapse();
-            if (!UseLocal)
+            if (!UseLocalPackagesFolder)
             {
-                string globalPackagesFolder = SettingsUtility.GetGlobalPackagesFolder(Settings);
-                if(!string.IsNullOrEmpty(globalPackagesFolder))
+                string globalPackagesFolder = SettingsUtility.GetGlobalPackagesFolder(_settings);
+                if (!string.IsNullOrEmpty(globalPackagesFolder))
                 {
                     packagesPath = packagesPath.Combine(new DirectoryPath(globalPackagesFolder)).Collapse();
                 }
             }
             return packagesPath;
         }
-
-        // Note that sources are searched first at index 0, then index 1, and so on until a match is found
-        public void AddPackageSource(string packageSource) => _packageSources.Insert(0, new PackageSource(packageSource));
-
-        public void AddPackage(string packageId, IReadOnlyList<string> packageSources, string versionSpec, 
-            bool getLatest, bool allowPrereleaseVersions, bool allowUnlisted, bool exclusive) 
-            => _packages.Add(new Package(packageId, packageSources, versionSpec, getLatest, allowPrereleaseVersions, allowUnlisted, exclusive));
-
-        // Based primarily on NuGet.CommandLine.Commands.UpdateCommand
+        
         internal void InstallPackages()
         {
-            Trace.Verbose($"Installing packages to {GetAbsolutePackagesPath().FullPath} (using {(UseLocal ? "local" : "global")} packages folder)");
-            SourceRepository localSourceRepository = SourceRepositoryProvider.CreateRepository(new PackageSource(GetAbsolutePackagesPath().FullPath));
-            List<SourceRepository> defaultSourceRepositories = _packageSources.Select(SourceRepositoryProvider.CreateRepository).ToList();
+            DirectoryPath packagesPath = GetAbsolutePackagesPath();
+            Trace.Verbose($"Installing packages to {packagesPath.FullPath} (using {(UseLocalPackagesFolder ? "local" : "global")} packages folder)");
+
+            // Add the global default sources if requested
+            if (UseGlobalPackageSources)
+            {
+                _sourceRepositories.AddGlobalDefaults();
+            }
+
+            // Get the local repository
+            SourceRepository localRepository = _sourceRepositories.CreateRepository(packagesPath.FullPath);
 
             // Install the packages
-            _nuGetProject.StartPackageInstall();
-            Parallel.ForEach(_packages, package =>
+            WyamFolderNuGetProject nuGetProject = new WyamFolderNuGetProject(_fileSystem, _assemblyLoader, _currentFramework, packagesPath.FullPath);
+            NuGetPackageManager packageManager = new NuGetPackageManager(_sourceRepositories, _settings, packagesPath.FullPath)
             {
-                List<SourceRepository> sourceRepositories = defaultSourceRepositories;
-                if (package.PackageSources != null && package.PackageSources.Count > 0)
-                {
-                    sourceRepositories = package.PackageSources.Select(SourceRepositoryProvider.CreateRepository).ToList();
-                    if (!package.Exclusive)
-                    {
-                        sourceRepositories = sourceRepositories.Concat(defaultSourceRepositories).ToList();
-                    }
-                }
-                package.InstallPackage(this, UpdatePackages, localSourceRepository, sourceRepositories);
-            });
+                PackagesFolderNuGetProject = nuGetProject
+            };
+            Task.WaitAll(_packages.Select(x => x.Install(localRepository, _sourceRepositories.GetDefaultRepositories(), UpdatePackages, packageManager, _logger)).ToArray());
 
             // Process the package (do this after all packages have been installed)
-            _nuGetProject.ProcessAssembliesAndContent();
+            nuGetProject.ProcessAssembliesAndContent();
         }
     }
 }